consolidate resources to DynamicEmergency namespace
