+++ lib/resources/DynamicEmergency.js
@@ -1,0 +1,10 @@
+'use strict';
+
+var TelnyxResource = require('../TelnyxResource');
+
+module.exports = TelnyxResource.extend({
+  nestedResources: {
+    Addresses: require('./DynamicEmergencyAddresses'),
+    Endpoints: require('./DynamicEmergencyEndpoints'),
+  }
+});
