+++ lib/telnyx.js
@@ -56,7 +56,7 @@ var resources = {
   Fqdns: require('./resources/Fqdns'),
   Connections: require('./resources/Connections'),
   IpConnections: require('./resources/IpConnections'),
-  DynamicEmergencyEndpoints: require('./resources/DynamicEmergencyEndpoints'),
+  DynamicEmergency: require('./resources/DynamicEmergency'),
   FqdnConnections: require('./resources/FqdnConnections'),
   CredentialConnections: require('./resources/CredentialConnections'),
   RegulatoryRequirements: require('./resources/RegulatoryRequirements'),
@@ -64,7 +64,6 @@ var resources = {
   NumberOrderDocuments: require('./resources/NumberOrderDocuments'),
   Actions: require('./resources/Actions'),
   OutboundVoiceProfiles: require('./resources/Outbound'),
-  DynamicEmergencyAddresses: require('./resources/DynamicEmergencyAddresses'),
   CallControlApplications: require('./resources/CallControlApplications'),
   PhoneNumbersInboundChannels: require('./resources/PhoneNumbersInboundChannels'),
   OtaUpdates: require('./resources/OtaUpdates'),
